+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/BenchmarkFunctions.hs
@@ -343,7 +343,6 @@ mkStakePoolParams keys =
     , sppOwners = Set.singleton $ hashKey (vKey stakeKeyOne)
     , sppRelays = StrictSeq.empty
     , sppMetadata = SNothing
-    , sppBlsKey = SNothing
     }
 
 -- Create stake pool registration certs