+++ libs/cardano-ledger-core/src/Cardano/Ledger/State/StakePool.hs
@@ -33,6 +33,9 @@ module Cardano.Ledger.State.StakePool (
   -- * Lenses
   spsVrfL,
   spsBlsKeyL,
+  BlsKeyState (..),
+  bksKeyL,
+  bksRegisteredInL,
   spsPledgeL,
   spsCostL,
   spsMarginL,
@@ -45,9 +48,6 @@ module Cardano.Ledger.State.StakePool (
 
   -- * Conversions
   mkStakePoolState,
-  BlsKeyState (..),
-  bksKeyL,
-  bksRegisteredInL,
   stakePoolStateToStakePoolParams,
 
   -- * Pool Parameters and Related Types
@@ -59,7 +59,6 @@ module Cardano.Ledger.State.StakePool (
     PoolParams,
     ppId,
     ppVrf,
-    ppBlsKey,
     ppPledge,
     ppCost,
     ppMargin,
@@ -104,9 +103,7 @@ import Cardano.Ledger.Binary (
   EncCBOR (..),
   Encoding,
   Interns,
-  TokenType (TypeListLen, TypeListLen64, TypeListLenIndef, TypeNull),
   decodeFixedSized,
-  decodeNull,
   decodeNullStrictMaybe,
   decodeRecordNamed,
   decodeRecordNamedT,
@@ -114,12 +111,8 @@ import Cardano.Ledger.Binary (
   encodeFixedSized,
   encodeListLen,
   encodeNullStrictMaybe,
-  ifDecoderVersionAtLeast,
-  natVersion,
-  peekTokenType,
   rawDecodeFixedSized,
   rawEncodeFixedSized,
-  withCurrentEncodingVersion,
  )
 import Cardano.Ledger.Binary.Coders (
   Decode (..),
@@ -163,6 +156,7 @@ data StakePoolState = StakePoolState
   { spsVrf :: !(VRFVerKeyHash StakePoolVRF)
   -- ^ VRF verification key hash for leader election
   , spsBlsKey :: !(StrictMaybe BlsKeyState)
+  -- ^ BLS verificaton key (and proof of possession) for voting.
   , spsPledge :: !Coin
   -- ^ Pledge amount committed by the pool operator
   , spsCost :: !Coin
@@ -284,15 +278,13 @@ instance Default StakePoolState where
 
 -- | Convert 'StakePoolParams' to 'StakePoolState' by dropping the pool ID.
 -- This is the primary way to create a 'StakePoolState' from registration
--- or update parameters. The epoch is the one the parameters take effect in:
--- a voting key carried by the parameters is stamped with it, so registering
--- again renews the key.
+-- or update parameters.
 mkStakePoolState ::
-  EpochNo -> CompactForm Coin -> Set (Credential Staking) -> StakePoolParams era -> StakePoolState
-mkStakePoolState epochNo deposit delegators spp =
+  CompactForm Coin -> Set (Credential Staking) -> StakePoolParams era -> StakePoolState
+mkStakePoolState deposit delegators spp =
   StakePoolState
     { spsVrf = sppVrf spp
-    , spsBlsKey = (`BlsKeyState` epochNo) <$> sppBlsKey spp
+    , spsBlsKey = SNothing
     , spsPledge = sppPledge spp
     , spsCost = sppCost spp
     , spsMargin = sppMargin spp
@@ -313,7 +305,6 @@ stakePoolStateToStakePoolParams networkId poolId sps =
   StakePoolParams
     { sppId = poolId
     , sppVrf = spsVrf sps
-    , sppBlsKey = bksKey <$> spsBlsKey sps
     , sppPledge = spsPledge sps
     , sppCost = spsCost sps
     , sppMargin = spsMargin sps
@@ -461,7 +452,6 @@ instance DecCBOR StakePoolRelay where
 data StakePoolParams era = StakePoolParams
   { sppId :: !(KeyHash StakePool)
   , sppVrf :: !(VRFVerKeyHash StakePoolVRF)
-  , sppBlsKey :: !(StrictMaybe BlsKey)
   , sppPledge :: !Coin
   , sppCost :: !Coin
   , sppMargin :: !UnitInterval
@@ -526,9 +516,9 @@ instance DecCBOR BlsKey where
     blsPossessionProof <- decodeFixedSized
     pure BlsKey {blsPubKey, blsPossessionProof}
 
--- | A registered voting key, together with the epoch its registration took
--- effect in. The key stops being honoured @maxKeyAgeEpochs@ epochs later, at
--- which point the pool's committee seat is keyless until it registers again.
+-- | A registered voting key, together with the epoch its registration certificate was
+-- accepted in. The key stops being honoured @maxKeyAgeEpochs@ epochs later, at which
+-- point the pool's committee seat is keyless until it registers again.
 data BlsKeyState = BlsKeyState
   { bksKey :: !BlsKey
   , bksRegisteredIn :: !EpochNo
@@ -565,7 +555,7 @@ sppMetadataL :: Lens' (StakePoolParams era) (StrictMaybe PoolMetadata)
 sppMetadataL = lens sppMetadata (\spp u -> spp {sppMetadata = u})
 
 instance Default (StakePoolParams era) where
-  def = StakePoolParams def def def (Coin 0) (Coin 0) def def def def def
+  def = StakePoolParams def def (Coin 0) (Coin 0) def def def def def
 
 instance NoThunks (StakePoolParams era)
 
@@ -576,7 +566,6 @@ instance ToJSON (StakePoolParams era) where
     Aeson.object
       [ "poolId" .= sppId spp
       , "vrf" .= sppVrf spp
-      , "blsKey" .= sppBlsKey spp
       , "pledge" .= sppPledge spp
       , "cost" .= sppCost spp
       , "margin" .= sppMargin spp
@@ -594,7 +583,6 @@ instance FromJSON (StakePoolParams era) where
       StakePoolParams
         <$> ((obj .: "poolId") <|> (obj .: "publicKey"))
         <*> obj .: "vrf"
-        <*> obj .:? "blsKey" .!= SNothing
         <*> obj .: "pledge"
         <*> obj .: "cost"
         <*> obj .: "margin"
@@ -608,7 +596,6 @@ type PoolParams = StakePoolParams
 pattern PoolParams ::
   KeyHash StakePool ->
   VRFVerKeyHash StakePoolVRF ->
-  StrictMaybe BlsKey ->
   Coin ->
   Coin ->
   UnitInterval ->
@@ -620,7 +607,6 @@ pattern PoolParams ::
 pattern PoolParams
   { ppId
   , ppVrf
-  , ppBlsKey
   , ppPledge
   , ppCost
   , ppMargin
@@ -632,7 +618,6 @@ pattern PoolParams
   StakePoolParams
     ppId
     ppVrf
-    ppBlsKey
     ppPledge
     ppCost
     ppMargin
@@ -648,7 +633,6 @@ pattern PoolParams
 {-# DEPRECATED
   ppId
   , ppVrf
-  , ppBlsKey
   , ppPledge
   , ppCost
   , ppMargin
@@ -705,24 +689,16 @@ withStakePoolParamsFlatEncoding ::
   (Int -> Encoding -> Encoding) ->
   Encoding
 withStakePoolParamsFlatEncoding poolParams f =
-  withCurrentEncodingVersion $ \v -> do
-    let (extraLen, blsKeyEncoding)
-          | v >= natVersion @12 =
-              case sppBlsKey poolParams of
-                SJust lk -> (1, encCBOR lk)
-                SNothing -> (0, mempty)
-          | otherwise = (0, mempty)
-    f (9 + extraLen) $
-      encCBOR (sppId poolParams)
-        <> encCBOR (sppVrf poolParams)
-        <> blsKeyEncoding
-        <> encCBOR (sppPledge poolParams)
-        <> encCBOR (sppCost poolParams)
-        <> encCBOR (sppMargin poolParams)
-        <> encCBOR (sppAccountAddress poolParams)
-        <> encCBOR (sppOwners poolParams)
-        <> encCBOR (sppRelays poolParams)
-        <> encodeNullStrictMaybe encCBOR (sppMetadata poolParams)
+  f 9 $
+    encCBOR (sppId poolParams)
+      <> encCBOR (sppVrf poolParams)
+      <> encCBOR (sppPledge poolParams)
+      <> encCBOR (sppCost poolParams)
+      <> encCBOR (sppMargin poolParams)
+      <> encCBOR (sppAccountAddress poolParams)
+      <> encCBOR (sppOwners poolParams)
+      <> encCBOR (sppRelays poolParams)
+      <> encodeNullStrictMaybe encCBOR (sppMetadata poolParams)
 
 instance Era era => DecCBOR (StakePoolParams era) where
   decCBOR = do
@@ -732,11 +708,6 @@ decodeStakePoolParamsFlat :: Decoder s (Int, StakePoolParams era)
 decodeStakePoolParamsFlat = do
   sppId <- decCBOR
   sppVrf <- decCBOR
-  (extraBlsKeyLen, sppBlsKey) <-
-    ifDecoderVersionAtLeast
-      (natVersion @12)
-      decodeBlsKeyBackwardsCompatible
-      (pure (0, SNothing))
   sppPledge <- decCBOR
   sppCost <- decCBOR
   sppMargin <- decCBOR
@@ -744,14 +715,4 @@ decodeStakePoolParamsFlat = do
   sppOwners <- decCBOR
   sppRelays <- decCBOR
   sppMetadata <- decodeNullStrictMaybe decCBOR
-  pure (9 + extraBlsKeyLen, StakePoolParams {..})
-  where
-    decodeBlsKeyBackwardsCompatible = do
-      peekTokenType >>= \case
-        TypeListLen -> (\lk -> (1, SJust lk)) <$> decCBOR
-        TypeListLen64 -> (\lk -> (1, SJust lk)) <$> decCBOR
-        TypeListLenIndef -> (\lk -> (1, SJust lk)) <$> decCBOR
-        TypeNull -> do
-          decodeNull
-          pure (1, SNothing)
-        _ -> pure (0, SNothing)
+  pure (9, StakePoolParams {..})