+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Pool.hs
@@ -71,7 +71,7 @@ poolCertSpec univ (PoolEnv e pp) ps =
     (caseOn pc)
       -- RegPool !(PoolParams c)
       ( branchW 1 $ \poolParams ->
-          match poolParams $ \_ _ _ _ cost _ accountAddr _ _ mMetadata ->
+          match poolParams $ \_ _ _ cost _ accountAddr _ _ mMetadata ->
             [ witness univ poolParams
             , match accountAddr $ \net' _ ->
                 net' ==. lit Testnet